sync

+++ keepalived/core/pidfile.c
@@ -20,25 +20,45 @@
  * Copyright (C) 2001-2012 Alexandre Cassen, <acassen@gmail.com>
  */
 
+#include "config.h"
+
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
 #include "logger.h"
 #include "pidfile.h"
-extern char *main_pidfile;
-extern char *checkers_pidfile;
-extern char *vrrp_pidfile;
+#include "main.h"
+#include "bitops.h"
 
-/* Create the runnnig daemon pidfile */
+const char *pid_directory = PID_DIR PACKAGE;
+
+/* Create the directory for non-standard pid files */
+void
+create_pid_dir(void)
+{
+	if (mkdir(pid_directory, S_IRWXU | S_IRGRP | S_IXGRP | S_IROTH | S_IXOTH) && errno != EEXIST) {
+		log_message(LOG_INFO, "Unable to create directory %s", pid_directory);
+		return;
+	}
+}
+
+void
+remove_pid_dir(void)
+{
+	if (rmdir(pid_directory) && errno != ENOTEMPTY && errno != EBUSY)
+		log_message(LOG_INFO, "unlink of %s failed - error (%d) '%s'", pid_directory, errno, strerror(errno));
+}
+
+/* Create the running daemon pidfile */
 int
-pidfile_write(char *pid_file, int pid)
+pidfile_write(const char *pid_file, int pid)
 {
 	FILE *pidfile = NULL;
 	int pidfd = creat(pid_file, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
 	if (pidfd != -1) pidfile = fdopen(pidfd, "w");
 
 	if (!pidfile) {
-		log_message(LOG_INFO, "pidfile_write : Can not open %s pidfile",
+		log_message(LOG_INFO, "pidfile_write : Cannot open %s pidfile",
 		       pid_file);
 		return 0;
 	}
@@ -49,17 +69,17 @@ pidfile_write(char *pid_file, int pid)
 
 /* Remove the running daemon pidfile */
 void
-pidfile_rm(char *pid_file)
+pidfile_rm(const char *pid_file)
 {
 	unlink(pid_file);
 }
 
 /* return the daemon running state */
-int
-process_running(char *pid_file)
+static int
+process_running(const char *pid_file)
 {
 	FILE *pidfile = fopen(pid_file, "r");
-	pid_t pid;
+	pid_t pid = 0;
 	int ret;
 
 	/* No pidfile */
@@ -67,10 +87,16 @@ process_running(char *pid_file)
 		return 0;
 
 	ret = fscanf(pidfile, "%d", &pid);
-	if (ret == EOF && ferror(pidfile) != 0) {
-		log_message(LOG_INFO, "Error opening pid file %s", pid_file);
-	}
 	fclose(pidfile);
+	if (ret != 1) {
+		log_message(LOG_INFO, "Error reading pid file %s", pid_file);
+		pid = 0;
+		pidfile_rm(pid_file);
+	}
+
+	/* What should we return - we don't know if it is running or not. */
+	if (!pid)
+		return 1;
 
 	/* If no process is attached to pidfile, remove it */
 	if (kill(pid, 0)) {
@@ -83,17 +109,18 @@ process_running(char *pid_file)
 }
 
 /* Return parent process daemon state */
-int
-keepalived_running(int mode)
+bool
+keepalived_running(unsigned long mode)
 {
 	if (process_running(main_pidfile))
-		return 1;
-	else if (mode & 1 || mode & 2)
-		return process_running((mode & 1) ? vrrp_pidfile :
-				       checkers_pidfile);
-
-	if (process_running(vrrp_pidfile) ||
-	    process_running(checkers_pidfile))
-		return 1;
-	return 0;
+		return true;
+#ifdef _WITH_VRRP_
+	if (__test_bit(DAEMON_VRRP, &mode) && process_running(vrrp_pidfile))
+		return true;
+#endif
+#ifdef _WITH_LVS_
+	if (__test_bit(DAEMON_CHECKERS, &mode) && process_running(checkers_pidfile))
+		return true;
+#endif
+	return false;
 }
